Move `merge_else_if`
