+++ crates/swc_ecma_minifier/src/compress/pure/mod.rs
@@ -459,6 +459,8 @@ impl VisitMut for Pure<'_> {
         s.visit_mut_children_with(self);
 
         self.optimize_expr_in_bool_ctx(&mut s.test, false);
+
+        self.merge_else_if(s);
     }
 
     fn visit_mut_member_expr(&mut self, e: &mut MemberExpr) {
